more corrections and add a config v2
