+++ crates/skytem_logs/src/mbed_motor_control/mbed_config.rs
@@ -2,8 +2,8 @@ use std::io::{self};
 
 pub(crate) mod v1;
 pub(crate) use v1::MbedConfigV1;
-pub(crate) mod v2;
-pub(crate) use v2::MbedConfigV2;
+pub(crate) mod v3;
+pub(crate) use v3::MbedConfigV3;
 
 pub(crate) trait MbedConfig: Sized {
     /// As long as implementors are packed structs, this can just return `size_of::<Self>()`, otherwise it should return
