+++ crates/skytem_logs/src/mbed_motor_control/pid/header/v4.rs
@@ -1,7 +1,7 @@
 use std::{fmt, io};
 
 use crate::mbed_motor_control::{
-    mbed_config::MbedConfigV2,
+    mbed_config::MbedConfigV3,
     mbed_header::{
         BuildMbedLogHeaderV2, GitBranchData, GitRepoStatusData, GitShortShaData,
         MbedMotorControlLogHeader, ProjectVersionData, StartupTimestamp, UniqueDescriptionData,
@@ -23,16 +23,16 @@ pub(crate) struct PidLogHeaderV4 {
     git_branch: GitBranchData,
     git_repo_status: GitRepoStatusData,
     startup_timestamp: StartupTimestamp,
-    mbed_config: MbedConfigV2,
+    mbed_config: MbedConfigV3,
 }
 
 impl PidLogHeaderV4 {
-    pub(crate) fn mbed_config(&self) -> &MbedConfigV2 {
+    pub(crate) fn mbed_config(&self) -> &MbedConfigV3 {
         &self.mbed_config
     }
 }
 
-impl BuildMbedLogHeaderV2<MbedConfigV2> for PidLogHeaderV4 {
+impl BuildMbedLogHeaderV2<MbedConfigV3> for PidLogHeaderV4 {
     fn new(
         unique_description: UniqueDescriptionData,
         version: u16,
@@ -41,7 +41,7 @@ impl BuildMbedLogHeaderV2<MbedConfigV2> for PidLogHeaderV4 {
         git_branch: GitBranchData,
         git_repo_status: GitRepoStatusData,
         startup_timestamp: StartupTimestamp,
-        mbed_config: MbedConfigV2,
+        mbed_config: MbedConfigV3,
     ) -> Self {
         Self {
             unique_description,
@@ -99,7 +99,7 @@ impl GitMetadata for PidLogHeaderV4 {
 }
 
 impl MbedMotorControlLogHeader for PidLogHeaderV4 {
-    const VERSION: u16 = 2;
+    const VERSION: u16 = 4;
 
     fn unique_description_bytes(&self) -> &UniqueDescriptionData {
         &self.unique_description
@@ -161,81 +161,6 @@ impl fmt::Display for PidLogHeaderV4 {
         if self.git_repo_status().is_some() {
             writeln!(f, "Repo status: dirty")?;
         }
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::fs;
-
-    const TEST_DATA: &str = "../../test_data/mbed_motor_control/v3/pid_20241023_143859_00.bin";
-
-    use io::Read;
-    use testresult::TestResult;
-
-    use super::*;
-
-    #[test]
-    fn test_deserialize() -> TestResult {
-        let mut file = fs::File::open(TEST_DATA)?;
-        let mut buffer = vec![0; 500];
-        file.read_exact(&mut buffer)?;
-        let (pid_log_header, bytes_read) = PidLogHeaderV4::from_reader(&mut buffer.as_slice())?;
-        eprintln!("{pid_log_header}");
-        assert_eq!(bytes_read, 329);
-        assert_eq!(
-            pid_log_header.unique_description(),
-            crate::mbed_motor_control::pid::UNIQUE_DESCRIPTION
-        );
-        assert_eq!(pid_log_header.version, 3);
-        assert_eq!(pid_log_header.project_version().unwrap(), "3.0.0");
-        assert_eq!(
-            pid_log_header.git_branch(),
-            Some("fix-23-pid-loop-in-standby".into())
-        );
-        assert_eq!(pid_log_header.git_short_sha(), Some("9057619".into()));
-        assert_eq!(pid_log_header.mbed_config().general_cfg().t_standby(), 50);
-        assert_eq!(pid_log_header.mbed_config().general_cfg().t_run(), 65);
-        assert_eq!(pid_log_header.mbed_config().general_cfg().t_fan_on(), 81);
-        assert_eq!(pid_log_header.mbed_config().general_cfg().t_fan_off(), 80);
-        assert_eq!(
-            pid_log_header.mbed_config().general_cfg().rpm_standby(),
-            3600
-        );
-        assert_eq!(
-            pid_log_header.mbed_config().general_cfg().rpm_running(),
-            6000
-        );
-        assert_eq!(
-            pid_log_header.mbed_config().general_cfg().time_shutdown(),
-            60
-        );
-        assert_eq!(
-            pid_log_header
-                .mbed_config()
-                .general_cfg()
-                .time_wait_for_cap(),
-            300
-        );
-        assert_eq!(
-            pid_log_header.mbed_config().general_cfg().vbat_ready(),
-            12.8
-        );
-        assert_eq!(pid_log_header.mbed_config().general_cfg().servo_max(), 1500);
-        assert_eq!(pid_log_header.mbed_config().general_cfg().servo_min(), 800);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().kp_initial(), 3.0);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().ki_initial(), 0.0);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().kd_initial(), 0.0);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().kp_idle(), 0.0);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().ki_idle(), 0.0);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().kd_idle(), 0.0);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().kp_standby(), 1.0);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().ki_standby(), 1.0);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().kd_standby(), 0.0);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().kp_running(), 3.0);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().ki_running(), 1.0);
-        assert_eq!(pid_log_header.mbed_config().pid_cfg().kd_running(), 0.0);
         Ok(())
     }
 }
